# apps/web/public/openapi.yaml
openapi: 3.0.0
info:
  title: VexonAC API
  description: |
    Official API documentation for the VexonAC Anti-Cheat System and its network tools.
    
    ## Authentication
    All API endpoints require authentication using API Keys.
    headers:
    - `x-api-key`: Your public Access Key
    - `x-api-secret`: Your private Secret Key
    
    ## Rate Limiting & Cost
    - Limits are applied per minute based on your API Key tier.
    - Each endpoint has a cost associated with it. Your API Key tier determines your monthly quota.
    - Response headers `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`, `X-Cost` included.
  version: 1.0.0
servers:
  - url: https://api.vexonac.xyz
    description: Production Server
security:
  - ApiKeyAuth: []
    ApiSecretAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
    ApiSecretAuth:
      type: apiKey
      in: header
      name: x-api-secret
  schemas:
...
    BanPublic:
      type: object
      properties:
        id: { type: string }
        player: { type: string, description: "Player Name" }
        server: { type: string, description: "Server Name" }
        reason: { type: string }
        details: { type: object }
        evidence: { type: string }
        bannedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        bannedBy: { type: string }
    ServerBanHistory:
      type: object
      properties:
        player: { type: string }
        reason: { type: string }
        details: { type: object }
        evidence: { type: string }
        bannedAt: { type: string, format: date-time }
...
    AnalysisBan:
      type: object
      properties:
        id: { type: string }
        banId: { type: string }
        reason: { type: string }
        details: { type: object }
        evidenceUrl: { type: string }
        bannedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        bannedBy: { type: string }
        server:
          type: object
          properties:
            id: { type: string }
            name: { type: string }
            isCurrentServer: { type: boolean }
        identifiers:
          type: array
          items:
            type: object
            properties:
              type: { type: string }
              value: { type: string }
        isDirectBan: { type: boolean }
        isAltAccountBan: { type: boolean }
        matchingIdentifiers: 
          type: array
          items: 
             type: object
             properties:
               type: { type: string }
               value: { type: string }
        player:
          type: object 
          properties:
             id: { type: string }
             playerName: { type: string }
             playTime: { type: integer }
    PlayerAnalysis:
      type: object
      properties:
        player:
          type: object
          properties:
            id: { type: string }
            playerName: { type: string }
            playerLicense: { type: string }
            identifiers: { type: array, items: { type: string } }
            oldIdentifiers: { type: array, items: { type: string } }
            firstJoin: { type: string, format: date-time }
            lastJoin: { type: string, format: date-time }
            playTime: { type: integer }
        altAccounts:
          type: array
          items:
            type: object
            properties:
              id: { type: string }
              playerName: { type: string }
              playerLicense: { type: string }
              firstJoin: { type: string, format: date-time }
              lastJoin: { type: string, format: date-time }
              playTime: { type: integer }
              server:
                type: object
                properties:
                  id: { type: string }
                  name: { type: string }
                  isCurrentServer: { type: boolean }
              matchingIdentifiers: { type: array, items: { type: string } }
              identifiersCount: { type: integer }
        playerBans:
          type: array
          items: { $ref: '#/components/schemas/AnalysisBan' }
        sameServerAltBans:
          type: array
          items: { $ref: '#/components/schemas/AnalysisBan' }
        crossServerBans:
          type: array
          items: { $ref: '#/components/schemas/AnalysisBan' }
        bansHistory:
          type: array
          items:
            type: object
            properties:
              server: { type: string }
              reason: { type: string }
              details: { type: object }
              evidenceUrl: { type: string }
              bannedAt: { type: string, format: date-time }
        threatScore:
          type: object
          properties:
            score: { type: number }
            riskLevel: { type: string, enum: [LOW, MEDIUM, HIGH, CRITICAL] }
            warnings: { type: array, items: { type: string } }
...
paths:
...
  /v1/player/search:
    get:
      tags: [Player]
      summary: Search Players
      description: Cost 5
      x-cost: 5
      parameters:
        - in: query
          name: name
          required: true
          schema: { type: string }
      responses:
        200:
          description: Results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        license: { type: string }
                        server: { type: string }
                        lastSeen: { type: string, format: date-time }
                        playTime: { type: integer }
  /v1/player/{identifier}:
    get:
      tags: [Player]
      summary: Player Profile
      description: Cost 5
      x-cost: 5
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: Profile
          content:
            application/json:
              schema:
                type: object
                properties:
                  name: { type: string }
                  license: { type: string }
                  stats:
                    type: object
                    properties:
                      globalPlayTime: { type: integer }
                      serversVisited: { type: integer }
                      firstSeen: { type: string, format: date-time }
                      lastSeen: { type: string, format: date-time }
                      lastServerJoined: { type: string }
                  bans:
                    type: object
                    properties:
                      activeBans: { type: integer }
                      totalBans: { type: integer }
  /v1/player/{identifier}/servers:
    get:
      tags: [Player]
      summary: Server History
      description: Cost 5
      x-cost: 5
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: History
          content:
            application/json:
              schema:
                type: object
                properties:
                  servers:
                    type: array
                    items:
                      type: object
                      properties:
                        id: { type: string }
                        name: { type: string }
                        firstJoin: { type: string, format: date-time }
                        lastJoin: { type: string, format: date-time }
                        playTime: { type: integer }
  /v1/player/{identifier}/bans:
    get:
      tags: [Player]
      summary: Player Active Bans
      description: Cost 5
      x-cost: 5
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: Active bans
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  bans:
                    type: array
                    items:
                      $ref: '#/components/schemas/BanPublic'
  /v1/player/{identifier}/bans-history:
    get:
      tags: [Player]
      summary: Player Ban History
      description: Cost 10
      x-cost: 10
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: Ban history
          content:
            application/json:
              schema:
                type: object
                properties:
                  count: { type: integer }
                  bans:
                    type: array
                    items:
                      $ref: '#/components/schemas/ServerBanHistory'
  /v1/player/{identifier}/threat-score:
    get:
      tags: [Player]
      summary: Threat Score
      description: Cost 5
      x-cost: 5
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: Score
          content:
            application/json:
              schema:
                type: object
                properties:
                  score: { type: number }
  /v1/player/{identifier}/analysis:
    get:
      tags: [Player]
      summary: Player Analysis
      description: |
        Performs a deep analysis of the player, including alt accounts, bans, and threat score.
        (Cost: 15)
      x-cost: 15
      parameters:
        - in: path
          name: identifier
          required: true
          schema: { type: string }
      responses:
        200:
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlayerAnalysis'

  # ===========================
  # WEBHOOKS
  # ===========================
...